backend/pyproject.toml
[project]
name = "lolo2"
version = "0.1.0"
description = ""
authors = [
    {name = "han2650",email = "wonhee2648@naver.com"}
]
requires-python = ">=3.10,<3.11"
dependencies = [
    "langchain (==0.3.27)",
    "langchain-core (==0.3.72)",
    "langchain-community (==0.3.27)",
    "langchain-openai (==0.3.28)",
    "langsmith (==0.4.8)",
    "openai (==1.97.1)",
    "pydantic (==2.11.7)",
    "python-dotenv (==1.1.1)",
    "sqlite-utils (==3.38)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "ipykernel (>=6.30.0,<7.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "notebook (>=7.4.4,<8.0.0)",
    "streamlit (>=1.48.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
